rlm/backend/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0


